TES-22: Refactor UI Tests
