+++ tests/test_ui/conftest.py
@@ -1,0 +1,9 @@
+import pytest
+from playwright.sync_api import sync_playwright
+
+@pytest.fixture(scope="module")
+def browser():
+    with sync_playwright() as p:
+        browser = p.chromium.launch()
+        yield browser
+        browser.close()
