+++ tests/test_ui/test_ui_drag_and_drop.py
@@ -1,12 +1,4 @@
 import pytest
-from playwright.sync_api import sync_playwright
-
-@pytest.fixture(scope="module")
-def browser():
-    with sync_playwright() as p:
-        browser = p.chromium.launch()
-        yield browser
-        browser.close()
 
 
 def test_ui_drag_and_drop(browser):
