+++ tests/test_ui/test_ui_limited_images.py
@@ -1,12 +1,4 @@
 import pytest
-from playwright.sync_api import sync_playwright
-
-@pytest.fixture(scope="module")
-def browser():
-    with sync_playwright() as p:
-        browser = p.chromium.launch()
-        yield browser
-        browser.close()
 
 
 def test_ui_with_limited_images(browser):
